order_api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pika==1.3.2
mysql-connector-python==8.0.33